Refactor polling code from testcase into more standalone interactor.
Working toward making interactor stand-alone and independent of the test framework.
